+++ packages/orq-rc/src/models/operations/createidentity.ts
@@ -139,7 +139,7 @@ export const CreateIdentityResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-03-12T06:24:07.289Z",
+    "2026-03-12T08:19:52.221Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {