+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -3599,7 +3599,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KKGBFE8QFVDY7KEKRT7B2ZWW"),
+  id: z.string().default("01KKGJ3BZSW72Y6B3M8Q68200S"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),