+++ packages/orq-rc/RELEASES.md
@@ -7649,3 +7649,13 @@ Based on:
 - [typescript v4.6.0-rc.3] packages/orq-rc
 ### Releases
 - [NPM v4.6.0-rc.3] https://www.npmjs.com/package/@orq-ai/node/v/4.6.0-rc.3 - packages/orq-rc
+
+## 2026-03-12 08:22:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.752.0 (2.859.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.6.0-rc.4] packages/orq-rc
+### Releases
+- [NPM v4.6.0-rc.4] https://www.npmjs.com/package/@orq-ai/node/v/4.6.0-rc.4 - packages/orq-rc